services/kafka_to_feature_store/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["jlozanol <jorchlozano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
quixstreams = "^3.1.0"
loguru = "^0.7.2"
hopsworks = "^4.1.0"
pydantic-settings = "^2.6.1"
python-dotenv = "^1.0.1"
pyarrow = "^18.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.7.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["I"]
